Moves the S3 sink bucket configuration up a level to simplify the YAML.
