+++ data-prepper-plugins/s3-sink/src/integrationTest/java/org/opensearch/dataprepper/plugins/sink/S3SinkServiceIT.java
@@ -5,19 +5,6 @@
 
 package org.opensearch.dataprepper.plugins.sink;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.mockito.Mockito.lenient;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-import java.time.Duration;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.DistributionSummary;
 import org.hamcrest.CoreMatchers;
@@ -38,7 +25,6 @@ import org.opensearch.dataprepper.plugins.sink.accumulator.InMemoryBufferFactory
 import org.opensearch.dataprepper.plugins.sink.accumulator.ObjectKey;
 import org.opensearch.dataprepper.plugins.sink.codec.JsonCodec;
 import org.opensearch.dataprepper.plugins.sink.configuration.AwsAuthenticationOptions;
-import org.opensearch.dataprepper.plugins.sink.configuration.BucketOptions;
 import org.opensearch.dataprepper.plugins.sink.configuration.ObjectKeyOptions;
 import org.opensearch.dataprepper.plugins.sink.configuration.ThresholdOptions;
 import software.amazon.awssdk.core.ResponseBytes;
@@ -49,6 +35,21 @@ import software.amazon.awssdk.services.s3.model.GetObjectResponse;
 import software.amazon.awssdk.services.s3.model.ListObjectsRequest;
 import software.amazon.awssdk.services.s3.model.ListObjectsResponse;
 import software.amazon.awssdk.services.s3.model.S3Object;
+
+import java.time.Duration;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.mockito.Mockito.lenient;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class S3SinkServiceIT {
@@ -66,8 +67,6 @@ class S3SinkServiceIT {
     private AwsAuthenticationOptions awsAuthenticationOptions;
     @Mock
     private ObjectKeyOptions objectKeyOptions;
-    @Mock
-    private BucketOptions bucketOptions;
     @Mock
     private PluginMetrics pluginMetrics;
     @Mock
@@ -94,13 +93,12 @@ class S3SinkServiceIT {
 
         when(objectKeyOptions.getNamePattern()).thenReturn("elb-log-%{yyyy-MM-dd'T'hh-mm-ss}");
         when(objectKeyOptions.getPathPrefix()).thenReturn(PATH_PREFIX);
-        when(bucketOptions.getBucketName()).thenReturn(bucketName);
-        when(bucketOptions.getObjectKeyOptions()).thenReturn(objectKeyOptions);
+        when(s3SinkConfig.getBucketName()).thenReturn(bucketName);
+        when(s3SinkConfig.getObjectKeyOptions()).thenReturn(objectKeyOptions);
         when(thresholdOptions.getEventCount()).thenReturn(2);
         when(thresholdOptions.getMaximumSize()).thenReturn(ByteCount.parse("2mb"));
         when(thresholdOptions.getEventCollectTimeOut()).thenReturn(Duration.parse("PT3M"));
         when(s3SinkConfig.getThresholdOptions()).thenReturn(thresholdOptions);
-        when(s3SinkConfig.getBucketOptions()).thenReturn(bucketOptions);
         when(s3SinkConfig.getAwsAuthenticationOptions()).thenReturn(awsAuthenticationOptions);
         when(awsAuthenticationOptions.getAwsRegion()).thenReturn(Region.of(s3region));
