+++ data-prepper-plugins/s3-sink/src/test/java/org/opensearch/dataprepper/plugins/sink/S3SinkTest.java
@@ -5,15 +5,6 @@
 
 package org.opensearch.dataprepper.plugins.sink;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Collection;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -27,10 +18,20 @@ import org.opensearch.dataprepper.plugins.sink.accumulator.BufferTypeOptions;
 import org.opensearch.dataprepper.plugins.sink.codec.Codec;
 import org.opensearch.dataprepper.plugins.sink.codec.JsonCodec;
 import org.opensearch.dataprepper.plugins.sink.configuration.AwsAuthenticationOptions;
-import org.opensearch.dataprepper.plugins.sink.configuration.BucketOptions;
 import org.opensearch.dataprepper.plugins.sink.configuration.ObjectKeyOptions;
 import org.opensearch.dataprepper.plugins.sink.configuration.ThresholdOptions;
 import software.amazon.awssdk.regions.Region;
+
+import java.time.Duration;
+import java.util.ArrayList;
+import java.util.Collection;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 class S3SinkTest {
 
@@ -47,7 +48,6 @@ class S3SinkTest {
     private S3Sink s3Sink;
     private PluginSetting pluginSetting;
     private PluginFactory pluginFactory;
-    private BucketOptions bucketOptions;
 
     @BeforeEach
     void setUp() {
@@ -57,7 +57,6 @@ class S3SinkTest {
         AwsAuthenticationOptions awsAuthenticationOptions = mock(AwsAuthenticationOptions.class);
         Codec codec = mock(JsonCodec.class);
         ObjectKeyOptions objectKeyOptions = mock(ObjectKeyOptions.class);
-        bucketOptions = mock(BucketOptions.class);
         pluginSetting = mock(PluginSetting.class);
         PluginModel pluginModel = mock(PluginModel.class);
         pluginFactory = mock(PluginFactory.class);
@@ -75,8 +74,7 @@ class S3SinkTest {
         when(pluginFactory.loadPlugin(any(), any())).thenReturn(codec);
         when(pluginSetting.getName()).thenReturn(SINK_PLUGIN_NAME);
         when(pluginSetting.getPipelineName()).thenReturn(SINK_PIPELINE_NAME);
-        when(s3SinkConfig.getBucketOptions()).thenReturn(bucketOptions);
-        when(s3SinkConfig.getBucketOptions().getBucketName()).thenReturn(BUCKET_NAME);
+        when(s3SinkConfig.getBucketName()).thenReturn(BUCKET_NAME);
     }
 
     @Test
@@ -105,8 +103,7 @@ class S3SinkTest {
 
     @Test
     void test_doOutput_with_empty_records() {
-        when(s3SinkConfig.getBucketOptions()).thenReturn(bucketOptions);
-        when(s3SinkConfig.getBucketOptions().getBucketName()).thenReturn(BUCKET_NAME);
+        when(s3SinkConfig.getBucketName()).thenReturn(BUCKET_NAME);
         s3Sink = new S3Sink(pluginSetting, s3SinkConfig, pluginFactory);
         Assertions.assertNotNull(s3Sink);
         s3Sink.doInitialize();
